aidevs3/w4/s04e05/requirements.txt
openai
PyYAML
requests
wget
pymupdf
